requirements.txt
requests
beautifulsoup4
uuid
python-dotenv